Eliminate main.py file ambiguity for SonarCloud analysis
- Remove redundant ciris_engine/main.py and CIRISGUI/apps/ciris-api/main.py
- Extract runtime utilities to ciris_engine/utils/runtime_utils.py
- Keep only the root main.py as the single entrypoint
- Update SonarCloud configuration to include root directory
- Remove main.py exclusions since ambiguity is now resolved
- Fix test_main_entrypoint.py to work with new structure

Fixes SonarCloud coverage report error: "Cannot resolve the file path
'main.py' of the coverage report, ambiguity, the file exists in several 'source'."

The root main.py remains fully functional with all CLI options intact.
All tests pass including the updated entrypoint tests.
