+++ sonar-project.properties
@@ -9,17 +9,16 @@ sonar.organization=cirisai
 sonar.python.version=3.12
 
 # Path is relative to the sonar-project.properties file. Replace "\" by "/" on Windows.
-# Only include ciris_engine to avoid main.py ambiguity issues
-sonar.sources=ciris_engine/
+# Include both main.py and ciris_engine/ - no more main.py ambiguity!
+sonar.sources=.,ciris_engine/
 
 # Coverage configuration
 sonar.python.coverage.reportPaths=coverage.xml
 
-# Exclude other main.py files from analysis to resolve ambiguity
-# Also exclude non-PL/SQL files from PL/SQL analysis
-sonar.exclusions=main.py,CIRISGUI/**/*,ciris_engine/persistence/migrations/*.sql
+# Exclude CIRISGUI and non-PL/SQL files from PL/SQL analysis
+sonar.exclusions=CIRISGUI/**/*,ciris_engine/persistence/migrations/*.sql
 
-# Include additional Python files outside ciris_engine if needed
+# Include all Python files
 sonar.inclusions=**/*.py
 
 # Disable PL/SQL analysis for SQLite migration files
